# backend/src/types/Apartment.ts
export default interface Apartment {
  id: string;
  name:string
  projectName:string
  area:number
  floorNumber:number
  price:number
  propertyType:string
  createdAt: string;
  updatedAt: string;
}